Made it so that you can only resize the form while not playing the game. Done this by setting the ResizeMode to CanResize when the game is stopped, and to NoResize while the game is in progress.
The initialisation of emojisToGuess has been removed from its declaration, as it now changes dependent on the size of the form.

The call to SetUpGame has been removed from the initialising MainWindow method, as it is now called in ProcessStart if "(emojisToGuess * 2 != gameEmoji.Count)" as well as for TimeTextBlock_MouseDown.